pyproject.toml
[tool.poetry]
name = "dev01"
version = "0.1.0"
description = ""
authors = ["Chigryaev Stas <chigryaevstas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
amqp = "5.0.2"
asgiref = "3.3.1"
astroid = "2.4.2"
billiard = "3.6.3.0"
celery = "5.0.5"
click = "7.1.2"
click-didyoumean = "0.0.3"
click-plugins = "1.1.1"
click-repl = "0.1.6"
Django = "3.1.5"
django-js-asset = "1.2.2"
django-mptt = "0.11.0"
djangorestframework = "3.12.2"
isort = "5.7.0"
kombu = "5.0.2"
lazy-object-proxy = "1.4.3"
mccabe = "0.6.1"
prompt-toolkit = "3.0.10"
pyinstrument = "3.3.0"
pyinstrument-cext = "0.2.3"
pylint = "2.6.0"
pytz = "2020.5"
six = "1.15.0"
sqlparse = "0.4.1"
toml = "0.10.2"
vine = "5.0.0"
wcwidth = "0.2.5"
wrapt = "1.12.1"
djoser = "^2.1.0"
Pillow = "^8.1.2"
black = "^20.8b1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
